dag/requirements.txt
google-generativeai>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.2
